PSD-883 completed restructure with proper housing of narrow and local integration tests
