+++ package.json
@@ -13,7 +13,7 @@
     "build:css": "node-sass --output=app/static/stylesheets --output-style=expanded app/static/stylesheets",
     "build:css-watch": "node-sass --output=app/static/stylesheets --output-style=expanded --source-map true --watch app/static/stylesheets",
     "pretest": "npm run test:lint",
-    "test": "npm run test:unit",
+    "test": "jest --runInBand",
     "test:integration": "jest --testPathPattern=test/integration",
     "test:unit": "jest --testPathPattern=test/unit",
     "test:watch": "jest --coverage=false --onlyChanged --watch --runInBand",
