+++ test/integration/local/message-service.test.js
@@ -1,4 +1,4 @@
-const messageService = require('../../app/services/message-service')
+const messageService = require('../../../app/services/message-service')
 
 describe('message service', () => {
   test('smoke test', async () => {
