+++ test/integration/narrow/routes/claim/mine-type.test.js
@@ -1,10 +1,10 @@
-describe('Healthy test', () => {
+describe('MineType test', () => {
   let createServer
   let server
 
   beforeAll(async () => {
-    jest.mock('../../../app/services/message-service')
-    createServer = require('../../../app/server')
+    jest.mock('../../../../../app/services/message-service')
+    createServer = require('../../../../../app/server')
   })
 
   beforeEach(async () => {
@@ -12,10 +12,10 @@ describe('Healthy test', () => {
     await server.initialize()
   })
 
-  test('GET /healthy route returns 200', async () => {
+  test('GET /claim/mine-type route returns 200', async () => {
     const options = {
       method: 'GET',
-      url: '/healthy'
+      url: '/claim/mine-type'
     }
 
     const response = await server.inject(options)
